requirements.txt
face_recognition==1.3.0
opencv-python==4.11.0.86
numpy==2.2.4
scikit-learn==1.6.1
pillow==11.1.0
dlib>=19.7 